+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - New keywords completion inside namespaced maps. #2113
 - Pass current namespace aliases to cljfmt when range-formatting. #2129
-- bump clj-konto to `2025.09.24-20251021.151059-11` improving performance and fixing false-positives.
+- bump clj-kondo to `2025.09.24-20251021.151059-11` improving performance and fixing false-positives.
+- Bump cljfmt to `0.15.1` - adding support for vertical alignment.
 
 ## 2025.08.25-14.21.46