+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Support `squint` projects when having `squint.edn`. #2158
 - Support find definition of java inner classes (Foo$Bar). #2157
 - Bump babashka/fs to `0.5.28`.
+- Bump opentelemetry to `1.51.0`.
 
 ## 2025.08.25-14.21.46